Adaptations to make GTBikeV work again with RAGECOOP
The NON_INTERACTIVE_MODE has been revisited to remove menus
and notifications.
Added client events for player connections and local player
connection and disconnection.
Modified the weather custom event to transmit the wind
information and not only the weather.
Split CustomEventReceivedArgs in server and client to make
it possible to build custom events using introspection that
works both in a server resource and a client resource.
Added a new data type int[] to the custom events serialization
and a new abstract class CustomEventMessage to create custom
events with automatic serialization deserialization.

Signed-off-by: Makinolo <makinolo@gmail.com>