Implement relevant peds sync only
Option for only syncing relevant peds (usually bots, but could be
others). Fixed and adapted all the needed parts of the client
to work properly with GTBikeV. Provided new API functions to set
the relevant peds and to extract ped ids.
Reestructured the logger class to look more like the log in
GTBikeV.